AccountActions

Description

createAccount

showAccount

transferRandom

credit

debit

transfer

Documentation

createAccount :: Int -> IO Account

The createAccount function saves a new Customer Account. It returns a newTVarIO of the saved account, in Customer format.

createCustomer

Arguments

:: Int

-> IO Customer
createCustomer takes one argument of type Int

The createCustomer function creates a Customer account. It returns a new Customer record with a random name, the given account number and account balance.

showAccount

Arguments

:: Account

-> IO ()
showAccount takes one argument of type Account

The showAccount function displays the account in a table format It returns an empty IO action

transferRandom

Arguments

:: [Account]

-> IO ()
transferRandom takes one argument: a list of type Account

The transferRandom function selects two random accounts and a random amount to transfer between them. It performs a random delay after the action. It returns an empty IO action

credit

Arguments

:: Int

-> Account

-> STM ()
credit takes two arguments of type Int and Account

The credit function reads the balance of an account and credits it with the given amount. It writes the new balance to the Account TVar. It returns an empty STM action

debit

Arguments

:: Int

-> Account

-> STM ()
debit takes two arguments of type Int and Account

The debit function reads the balance of an account and debits the specified amount from the account. It writes the new balance to the Account TVar. It returns an empty STM action

transfer

Arguments

:: Int

-> Account

-> Account

-> STM ()
transfer takes three arguments of type Int, Account and Account

The transfer function debits one customer account, and credits the amount to another customer account. It makes use of the credit and debit functions. It uses atomically to ensure that actions are isolated. It returns an empty IO action