DataTypes

Description

Customer

Account

Documentation

data Customer

The Customer constructor creates a new record for a customer | Customer is a derivation of the Show and Generic instances

Constructors

Customer

Fields

name :: String

The customer name of type String

number :: Int

The account number of type Int

balance :: Int

The account balance of type Int

type Account = TVar Customer

The Account type is a TVar of type Customer